fixed homepage layout,  moved navbar to top and footer to bottom and renamed compnents to components
