+++ app/components/NavBar.jsx
