+++ app/home/page.js
@@ -1,11 +1,11 @@
 import Image from "next/image";
-import NavBar from "../compnents/NavBar";
-import Footer from "../compnents/Footer";
+import NavBar from "../components/NavBar";
+import Footer from "../components/Footer";
 
 export default function Home(){
     return(
 <>
-    <Footer/>
+    <NavBar />
 
     <div className="flex flex-col justify-center  items-start w-[700px] h-[700px] px-10 ml-120">
     
@@ -62,7 +62,7 @@ export default function Home(){
         </div>
     </div> 
 
-    <NavBar/>
+    <Footer />
 
 
 </>
