+++ app/page.js
@@ -1,0 +1,5 @@
+import Home from "./home/page";
+
+export default function Page() {
+  return <Home />;
+}
